backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.0
reportlab==4.0.7
weasyprint==60.1
pydantic==2.4.2
python-jose[cryptography]==3.3.0
httpx==0.25.1
aiofiles==23.2.1